requirements.txt
beautifulsoup4==4.11.1
Pillow==9.2.0
requests==2.28.1
